+++ packages/sharing-editor/public/samples/011_component_gallery/pages/text.markdown1.py
@@ -1,0 +1,16 @@
+import streamlit as st
+
+md = st.text_area(
+    "Type in your markdown string (without outer quotes)",
+    "Happy Streamlit-ing! :balloon:",
+)
+
+st.code(
+    f"""
+import streamlit as st
+        
+st.markdown('''{md}''')
+"""
+)
+
+st.markdown(md)