+++ packages/sharing-editor/public/samples/011_component_gallery/pages/widget.date_input_empty.py
@@ -1,0 +1,5 @@
+import datetime
+import streamlit as st
+
+d = st.date_input("When's your birthday", value=None)
+st.write("Your birthday is:", d)