+++ CHANGELOG.md
@@ -7,6 +7,68 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.45.0] - 2024-01-20
+
+### `@stlite/kernel`
+
+#### Changed
+
+- Update Pyodide to 0.24.1, [#681](https://github.com/whitphx/stlite/pull/681).
+
+## [0.44.0] - 2024-01-20
+
+### `@stlite/kernel`
+
+#### Changed
+
+- Update Streamlit to 1.30.0, [#678](https://github.com/whitphx/stlite/pull/678)
+
+## [0.43.1] - 2024-01-19
+
+### `@stlite/kernel`
+
+#### Changed
+
+- Update Streamlit to 1.27.0, [#671](https://github.com/whitphx/stlite/pull/671)
+
+### `@stlite/mountable`
+
+#### Added
+
+- `disableProgressToasts` and `disableErrorToasts` options, [#671](https://github.com/whitphx/stlite/pull/671)
+
+#### Changed
+
+- Remove `ctx.disableToast()` and `ctx.enableToast()`, [#671](https://github.com/whitphx/stlite/pull/671)
+
+## [0.43.0]
+
+Skipped
+
+## [0.42.5] - 2024-01-19
+
+### `@stlite/kernel`
+
+#### Fixed
+
+- Install the user-specified requirements and the stlite custom wheels at the same time, [#676](https://github.com/whitphx/stlite/pull/676).
+
+## [0.42.4] - 2023-12-27
+
+### `@stlite/kernel`
+
+#### Fixed
+
+- Set the `altair` version to `<5.2.0` as a workaround to solve an error of [#662](https://github.com/whitphx/stlite/issues/662), [#668](https://github.com/whitphx/stlite/pull/668).
+- Install the user-specified requirements before the Streamlit package, [#664](https://github.com/whitphx/stlite/pull/664).
+- Fix `DEVELOPMENT.md`, [#665](https://github.com/whitphx/stlite/pull/665), by @andeplane.
+
+### `@stlite/desktop`
+
+#### Fixed
+
+- Fix the CSP for `st.camera_input()`, [#666](https://github.com/whitphx/stlite/pull/666).
+
 ## [0.42.3] - 2023-11-01
 
 ### `@stlite/kernel`